refactor(frontend): ブログ記事をarXiv/PDFで別ルートに分割
ハイブリッドloaderを廃止し、paper_idの`pdf-`接頭でルートを分離。
- /blog/:paperId: arXiv(公開)専用。サーバーloaderでSSRしSEO/OGPを維持。
  pdf-接頭のIDは /blog/pdf/:paperId へリダイレクト
- /blog/pdf/:paperId: PDF(非公開)専用。clientLoaderでユーザーの認証付き取得。
  noindexでクローラー非公開
表示ロジックはBlogPostViewに抽出して両ルートで共有。リンク生成は
blogPostPathヘルパーでpaper_id種別に応じて振り分け。
