+++ frontend/app/components/figure-suggestion/figure-lightbox.tsx
@@ -9,6 +9,7 @@ import {
   DialogTitle,
 } from '~/components/ui/dialog'
 import type { FigureSuggestionItem } from '~/hooks/use-figure-suggestion'
+import { blogPostPath } from '~/lib/blog-path'
 
 type FigureLightboxProps = {
   item: FigureSuggestionItem | null
@@ -45,7 +46,7 @@ export function FigureLightbox({ item, onClose }: FigureLightboxProps) {
             )}
 
             <Link
-              to={`/blog/${item.paper_id}`}
+              to={blogPostPath(item.paper_id)}
               className="inline-flex w-fit items-center gap-1.5 text-sm font-medium text-primary underline-offset-4 hover:underline"
             >
               <ExternalLinkIcon className="size-3.5" />
